ai/playground/livekit-tracing/pyproject.toml
[project]
name = "livekit-tracing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "langfuse>=3.6.0",
    "livekit-agents[cartesia,google,openai]>=1.2.14",
    "opentelemetry-api>=1.37.0",
]
